Removes po_localizer & test, which are refactored.
